requirements/base.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
flask>=3.0,<4.0
pandas>=2.0,<3.0
numpy>=1.24,<3.0
requests>=2.31,<3.0
google-api-python-client>=2.120,<3.0
google-auth>=2.29,<3.0
pytest>=8.0,<9.0
mlflow>=2.10,<3.0
# Airflow is intentionally isolated in requirements/airflow.txt because
# apache-airflow 2.9.x requires Flask<2.3 while the monitoring demo uses Flask>=3.
